ProductGeometry -> CylindricalChannelGeometry; move CylindrycalChannelGeometry and DiffusionLimitedReaction into chemical_reactions
